Moving API calls to APIManager class
